# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsHub API",
        "description": "The NewsHub API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sources": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get all sources",
                "description": "Returns a paginated list of sources with optional filtering",
                "operationId": "3961c337f9d8a0b11f9e8c6e0c0f4254",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter sources by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Comma-separated relations to include (e.g., articles). Supports nested relations (e.g. articles.category)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Source"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sources/{id}": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get a specific source",
                "description": "Returns a single source by ID with optional relations",
                "operationId": "78d7f62e5c2412a71386f6adfdf09eb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the source",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Comma-separated relations to include (e.g., articles). Supports nested relations (e.g. articles.category)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Source"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "http://example.com/api/resource"
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "http://example.com/api/resource?page=2"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": "http://example.com/api/resource?page=1"
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://example.com/api/resource?page=3"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "from": {
                        "type": "integer",
                        "example": 16
                    },
                    "to": {
                        "type": "integer",
                        "example": 30
                    }
                },
                "type": "object"
            },
...
            "Source": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}